Moved rangers back to floats - that's what the should be. Sliderbar now renders itself.
git-svn-id: https://www.cs.binghamton.edu/svn/simple@296 bb5c8b44-a00f-0410-a998-867a3eb9c5a3

+++ simplegui/sg_dragable.h
@@ -35,7 +35,7 @@ public:
   
 protected:
 //  static GL_MODEL Default_Mouse_Cursor;
-  double base_x, base_y, start_x, start_y;
+  float base_x, base_y, start_x, start_y;
   };
 
 #endif // SG_DRAGABLE_H
